// src/app/model/booking.ts
export class Booking{
  id?: number;
  bookingId?:number;
  room_name?: string;
  room_price?: any;
  room_photo?: string;
  room_capacity?: number;
  firstName?: string;
  lastName?: string;
  email?: string;
  mobile?: string;
  checkIn?: Date;
  checkOut?: Date;
  adults?: number;
  children?: number;
  status?: string;
  numOfGuests?: number;
  total?: number
}